[Profiler] Rename utilization plot module with an underscore and refer to new name (#2505)
Making a quick fix on file name changes discussed in @pedropontesgarcia
's previous PR!
